+++ shuup/admin/modules/contacts/views/detail.py
@@ -11,16 +11,17 @@ import warnings
 
 from django.conf import settings
 from django.contrib import messages
+from django.contrib.auth import get_user_model
 from django.core.urlresolvers import reverse
 from django.http.response import HttpResponseRedirect
 from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
 from django.views.generic import DetailView
-from django.contrib.auth.models import User
 
 from shuup.admin.modules.contacts.utils import (
     check_contact_permission, request_limited
 )
+from shuup.admin.modules.users.views.detail import get_admin_url, get_front_url
 from shuup.admin.shop_provider import get_shop
 from shuup.admin.toolbar import (
     DropdownActionButton, DropdownDivider, PostActionButton, Toolbar,
@@ -29,7 +30,6 @@ from shuup.admin.toolbar import (
 from shuup.apps.provides import get_provide_objects
 from shuup.core.models import CompanyContact, Contact
 from shuup.front.apps.registration.signals import company_contact_activated
-from shuup.admin.modules.users.views.detail import get_admin_url, get_front_url
 from shuup.utils.deprecation import RemovedFromShuupWarning
 from shuup.utils.excs import Problem
 
@@ -140,7 +140,7 @@ class ContactDetailToolbar(Toolbar):
     def build_login_as_button(self):
         user = self.contact.user if hasattr(self.contact, "user") else None
         current_user = self.request.user
-        if isinstance(user, User):
+        if isinstance(user, get_user_model()):
             has_privileges = bool(
                 getattr(current_user, "is_superuser", False) or
                 getattr(current_user, "is_staff", False)