+++ shuup/admin/modules/menu/views/arrange.py
@@ -12,9 +12,8 @@ import json
 from django.contrib import messages
 from django.core.urlresolvers import reverse_lazy
 from django.http import HttpResponseRedirect
-from django.views.generic import RedirectView, TemplateView
-
 from django.utils.translation import ugettext_lazy as _
+from django.views.generic import RedirectView, TemplateView
 
 from shuup import configuration
 from shuup.admin.menu import get_menu_entry_categories