+++ shuup/core/basket/objects.py
@@ -7,9 +7,9 @@
 # LICENSE file in the root directory of this source tree.
 from __future__ import unicode_literals
 
-import random
 from collections import Counter
 from decimal import Decimal
+from uuid import uuid4
 
 import six
 from django.contrib import messages
@@ -24,6 +24,7 @@ from shuup.core.models import (
 from shuup.core.order_creator import OrderSource, SourceLine
 from shuup.core.order_creator._source import LineSource
 from shuup.core.pricing._context import PricingContext
+from shuup.utils.http import get_client_ip
 from shuup.utils.numbers import parse_decimal_string
 from shuup.utils.objects import compare_partial_dicts
 
@@ -117,7 +118,7 @@ class BaseBasket(OrderSource):
         self.basket_name = basket_name
         self.key = basket_name
         if request:
-            self.ip_address = request.META.get("REMOTE_ADDR")
+            self.ip_address = get_client_ip(request)
         self.storage = get_storage()
         self._data = None
         self._shipping_address = None
@@ -151,7 +152,7 @@ class BaseBasket(OrderSource):
             try:
                 self._data = self.storage.load(basket=self)
             except BasketCompatibilityError as error:
-                msg = _("Basket loading failed: Incompatible basket (%s)")
+                msg = _("Basket loading failed: Incompatible basket (%s).")
                 messages.error(self.request, msg % error)
                 self.storage.delete(basket=self)
                 self._data = self.storage.load(basket=self)
@@ -448,8 +449,7 @@ class BaseBasket(OrderSource):
             raise ValueError("Error! Add a variation parent to the basket is not allowed.")
 
         return {
-            # TODO: FIXME: Make sure line_id's are unique (not random)
-            "line_id": str(random.randint(0, 0x7FFFFFFF)),
+            "line_id": uuid4().hex,
             "product": product,
             "supplier": supplier,
             "shop": shop,