+++ shuup/front/views/product.py
@@ -7,8 +7,8 @@
 # LICENSE file in the root directory of this source tree.
 from django.core.urlresolvers import reverse
 from django.http.response import HttpResponseRedirect
-from django.utils.translation import get_language
 from django.utils.translation import ugettext as _
+from django.utils.translation import get_language
 from django.views.generic import DetailView
 
 from shuup.core.models import Product, ProductMode, ShopProduct, Supplier